Deleted old unfinished code under Framework.Communications that wasn't being used anywhere. Also folded GenericAsyncResult into RestClient, since it is used only there. This is preparation to remove Framework.Communications entirely.
