--- OpenSim/Framework/Communications/Limit/IRequestLimitStrategy.cs
@@ -1,66 +1,0 @@
-/*
- * Copyright (c) Contributors, http://opensimulator.org/
- * See CONTRIBUTORS.TXT for a full list of copyright holders.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
- *     * Redistributions of source code must retain the above copyright
- *       notice, this list of conditions and the following disclaimer.
- *     * Redistributions in binary form must reproduce the above copyright
- *       notice, this list of conditions and the following disclaimer in the
- *       documentation and/or other materials provided with the distribution.
- *     * Neither the name of the OpenSimulator Project nor the
- *       names of its contributors may be used to endorse or promote products
- *       derived from this software without specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY THE DEVELOPERS ``AS IS'' AND ANY
- * EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
- * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
- * DISCLAIMED. IN NO EVENT SHALL THE CONTRIBUTORS BE LIABLE FOR ANY
- * DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
- * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
- * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
- * ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
- * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-namespace OpenSim.Framework.Communications.Limit
-{
-    /// <summary>
-    /// Interface for strategies that can limit requests from the client.  Currently only used in the
-    /// texture modules to deal with repeated requests for certain textures.  However, limiting strategies
-    /// could be used with other requests.
-    /// </summary>
-    public interface IRequestLimitStrategy<TId>
-    {
-        /// <summary>
-        /// Should the request be allowed?  If the id is not monitored, then the request is always allowed.
-        /// Otherwise, the strategy criteria will be applied.
-        /// </summary>
-        /// <param name="id"></param>
-        /// <returns></returns>
-        bool AllowRequest(TId id);
-
-        /// <summary>
-        /// Has the request been refused just once?
-        /// </summary>
-        /// <returns>False if the request has not yet been refused, or if the request has been refused more
-        /// than once.</returns>
-        bool IsFirstRefusal(TId id);
-
-        /// <summary>
-        /// Start monitoring for future AllowRequest calls.  If the id is already monitored, then monitoring
-        /// continues.
-        /// </summary>
-        /// <param name="id"></param>
-        void MonitorRequests(TId id);
-
-        /// <summary>
-        /// Is the id being monitored?
-        /// </summary>
-        /// <param name="uuid"> </param>
-        /// <returns></returns>
-        bool IsMonitoringRequests(TId id);
-    }
-}
