--- OpenSim/Framework/Communications/XMPP/XmppMessageStanza.cs
@@ -1,93 +1,0 @@
-/*
- * Copyright (c) Contributors, http://opensimulator.org/
- * See CONTRIBUTORS.TXT for a full list of copyright holders.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
- *     * Redistributions of source code must retain the above copyright
- *       notice, this list of conditions and the following disclaimer.
- *     * Redistributions in binary form must reproduce the above copyright
- *       notice, this list of conditions and the following disclaimer in the
- *       documentation and/or other materials provided with the distribution.
- *     * Neither the name of the OpenSimulator Project nor the
- *       names of its contributors may be used to endorse or promote products
- *       derived from this software without specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY THE DEVELOPERS ``AS IS'' AND ANY
- * EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
- * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
- * DISCLAIMED. IN NO EVENT SHALL THE CONTRIBUTORS BE LIABLE FOR ANY
- * DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
- * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
- * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
- * ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
- * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-using System.Xml.Serialization;
-
-namespace OpenSim.Framework.Communications.XMPP
-{
-    /// <summary>
-    /// Message types.
-    /// </summary>
-    public enum XmppMessageType
-    {
-        [XmlEnum("chat")] chat,
-        [XmlEnum("error")] error,
-        [XmlEnum("groupchat")] groupchat,
-        [XmlEnum("headline")] headline,
-        [XmlEnum("normal")] normal,
-    }
-
-    /// <summary>
-    /// Message body.
-    /// </summary>
-    public class XmppMessageBody
-    {
-        [XmlText]
-        public string Text;
-
-        public XmppMessageBody()
-        {
-        }
-
-        public XmppMessageBody(string message)
-        {
-            Text = message;
-        }
-
-        new public string ToString()
-        {
-            return Text;
-        }
-    }
-
-    [XmlRoot("message")]
-    public class XmppMessageStanza: XmppStanza
-    {
-        /// <summary>
-        /// IQ type: one of set, get, result, error
-        /// </summary>
-        [XmlAttribute("type")]
-        public XmppMessageType MessageType;
-
-        // [XmlAttribute("error")]
-        // public XmppError Error;
-
-        [XmlElement("body")]
-        public XmppMessageBody Body;
-
-        public XmppMessageStanza() : base()
-        {
-        }
-
-        public XmppMessageStanza(string fromJid, string toJid, XmppMessageType mType, string message) :
-            base(fromJid, toJid)
-        {
-            MessageType = mType;
-            Body = new XmppMessageBody(message);
-        }
-    }
-}
