--- OpenSim/Framework/Communications/XMPP/XmppPresenceStanza.cs
@@ -1,69 +1,0 @@
-/*
- * Copyright (c) Contributors, http://opensimulator.org/
- * See CONTRIBUTORS.TXT for a full list of copyright holders.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
- *     * Redistributions of source code must retain the above copyright
- *       notice, this list of conditions and the following disclaimer.
- *     * Redistributions in binary form must reproduce the above copyright
- *       notice, this list of conditions and the following disclaimer in the
- *       documentation and/or other materials provided with the distribution.
- *     * Neither the name of the OpenSimulator Project nor the
- *       names of its contributors may be used to endorse or promote products
- *       derived from this software without specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY THE DEVELOPERS ``AS IS'' AND ANY
- * EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
- * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
- * DISCLAIMED. IN NO EVENT SHALL THE CONTRIBUTORS BE LIABLE FOR ANY
- * DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
- * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
- * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
- * ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
- * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-using System.Xml.Serialization;
-
-namespace OpenSim.Framework.Communications.XMPP
-{
-    /// <summary>
-    /// Message types.
-    /// </summary>
-    public enum XmppPresenceType
-    {
-        [XmlEnum("unavailable")]  unavailable,
-        [XmlEnum("subscribe")]    subscribe,
-        [XmlEnum("subscribed")]   subscribed,
-        [XmlEnum("unsubscribe")]  unsubscribe,
-        [XmlEnum("unsubscribed")] unsubscribed,
-        [XmlEnum("probe")]        probe,
-        [XmlEnum("error")]        error,
-    }
-
-
-    [XmlRoot("message")]
-    public class XmppPresenceStanza: XmppStanza
-    {
-        /// <summary>
-        /// IQ type: one of set, get, result, error
-        /// </summary>
-        [XmlAttribute("type")]
-        public XmppPresenceType PresenceType;
-
-        // [XmlAttribute("error")]
-        // public XmppError Error;
-
-        public XmppPresenceStanza() : base()
-        {
-        }
-
-        public XmppPresenceStanza(string fromJid, string toJid, XmppPresenceType pType) :
-            base(fromJid, toJid)
-        {
-            PresenceType = pType;
-        }
-    }
-}
